requirements.txt
    fastapi
    uvicorn
    motor
    python-dotenv
    pymongo[srv]
